requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
redis
pytest
python-json-logger
black
prometheus-fastapi-instrumentator
colorlog